Move spiral class into map directory
