+++ engine/src/main/kotlin/world/gregs/voidps/engine/map/Spiral.kt
@@ -1,8 +1,6 @@
-package world.gregs.voidps.engine.client.update.task.viewport
+package world.gregs.voidps.engine.map
 
 import world.gregs.voidps.engine.entity.character.player.Viewport.Companion.VIEW_RADIUS
-import world.gregs.voidps.engine.map.Delta
-import world.gregs.voidps.engine.map.Tile
 import world.gregs.voidps.engine.map.chunk.Chunk
 
 object Spiral {
