+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/SpecialAttackMelee.kt
@@ -1,6 +1,5 @@
 package world.gregs.voidps.world.interact.entity.player.combat.melee
 
-import world.gregs.voidps.engine.client.update.task.viewport.spiral
 import world.gregs.voidps.engine.entity.character.Character
 import world.gregs.voidps.engine.entity.character.npc.NPCs
 import world.gregs.voidps.engine.entity.character.player.Player
@@ -9,6 +8,7 @@ import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
+import world.gregs.voidps.engine.map.spiral
 import world.gregs.voidps.engine.utility.inject
 import world.gregs.voidps.world.interact.entity.combat.*
 import kotlin.math.floor
